+++ Editor/ComponentPopupWindow.cs
@@ -2,17 +2,15 @@
 
 using System;
 using UnityEngine;
-
+using System.Collections.Generic;
 #if UNITY_EDITOR
 using UnityEditor;
 #endif
 
 using HananokiEditor.Extensions;
 
-#if UNITY_EDITOR
-#endif
 
-
+using UnityObject = UnityEngine.Object;
 namespace HananokiEditor.CustomHierarchy {
 	public class ComponentPopupWindow : HEditorWindow {
 
@@ -25,7 +23,22 @@ namespace HananokiEditor.CustomHierarchy {
 		[NonSerialized]
 		public Editor editor;
 
+		static List<Type> _types;
+		List<Type> m_types {
+			get {
+				if( _types == null ) {
+					_types = new List<Type>();
+					_types.Add( UnityTypes.TMPro_TMP_Text );
+					_types.Add( UnityTypes.UnityEngine_UI_Image );
+					_types.Add( UnityTypes.UnityEngine_UI_RawImage );
+					_types.Add( UnityTypes.UnityEngine_SpriteRenderer );
+				}
+				return _types;
+			}
+		}
 
+		SerializedObject m_serializedObject;
+		SerializedProperty m_RaycastTarget;
 
 		public Vector2 GetWindowSize() {
 			//	var v2 = com.GetWindowSize();
@@ -44,12 +57,14 @@ namespace HananokiEditor.CustomHierarchy {
 			window.typeName = com.GetType().AssemblyQualifiedName;
 
 			window.closeEvent = closeEvent;
-			window.OnEnable2();
+			window.OnInitSerializedObject( com );
 		}
 
-		void OnEnable2() {
-
+		void OnInitSerializedObject( UnityObject obj ) {
+			m_serializedObject = new SerializedObject( obj );
+			m_RaycastTarget = m_serializedObject.FindProperty( "m_RaycastTarget" );
 		}
+
 		void OnSelectionChange() {
 			Component comp = null;
 			if( Selection.activeGameObject == null ) {
@@ -59,20 +74,20 @@ namespace HananokiEditor.CustomHierarchy {
 				}
 			}
 			else {
-				comp = (Component) Selection.activeGameObject.GetComponent( UnityTypes.TMPro_TMP_Text );
-				if( comp == null ) {
-					comp = (Component) Selection.activeGameObject.GetComponent( UnityTypes.UnityEngine_UI_Image );
-					if( comp == null ) {
-						comp = (Component) Selection.activeGameObject.GetComponent( UnityTypes.UnityEngine_SpriteRenderer );
-					}
+				foreach( var t in m_types ) {
+					comp = (Component) Selection.activeGameObject.GetComponent( t );
+					if( comp != null ) break;
 				}
 			}
+
 
 			//var comp = Selection.activeGameObject.GetComponent( type );
 			if( comp == null ) return;
 			com = comp;
 			typeName = com.GetType().AssemblyQualifiedName;
 			com = comp;
+			SetTitle( com.GetType().Name );
+			OnInitSerializedObject( comp );
 			Repaint();
 		}
 
@@ -88,6 +103,7 @@ namespace HananokiEditor.CustomHierarchy {
 
 			//editor.OnInspectorGUI();
 			if( UnityTypes.TMPro_TMP_Text != null && com.GetType().IsSubclassOf( UnityTypes.TMPro_TMP_Text ) ) {
+				//var _TMP_Text = (TMP_Text) com;
 				var s = com.GetProperty<string>( "text" );
 				ScopeHorizontal.Begin();
 				if( GUILayout.Button( "#", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "#" );
@@ -99,16 +115,15 @@ namespace HananokiEditor.CustomHierarchy {
 				s = EditorGUILayout.TextArea( s, EditorStyles.textField, GUILayout.Height( 20 * 3 ) );
 				if( ScopeChange.End() ) {
 					com.SetProperty<string>( "text", s );
-					SceneView.RepaintAll();
-					//Debug.Log( typeof( Image ).AssemblyQualifiedName );
-					//Debug.Log( typeof( RawImage ).AssemblyQualifiedName );
+					EditorWindowUtils.RepaintSceneView();
 				}
+
 			}
 			if( UnityTypes.UnityEngine_UI_Image != null && com.GetType() == ( UnityTypes.UnityEngine_UI_Image ) ) {
 				ScopeHorizontal.Begin();
 				if( GUILayout.Button( "#", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "#" );
-					if( GUILayout.Button( "##", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "##" );
-					if( GUILayout.Button( "Set GameObject Name" ) ) EditorHelper.SetGameObjectName( com );
+				if( GUILayout.Button( "##", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "##" );
+				if( GUILayout.Button( "Set GameObject Name" ) ) EditorHelper.SetGameObjectName( com );
 				ScopeHorizontal.End();
 
 				var s = com.GetProperty<Sprite>( "sprite" );
@@ -120,12 +135,24 @@ namespace HananokiEditor.CustomHierarchy {
 					//Debug.Log( typeof( Image ).AssemblyQualifiedName );
 					//Debug.Log( typeof( RawImage ).AssemblyQualifiedName );
 				}
+
 			}
+			if( m_RaycastTarget != null ) {
+				ScopeChange.Begin();
+				var b = HEditorGUILayout.ToggleLeft( "raycastTarget".nicify(), m_RaycastTarget.boolValue );
+				if( ScopeChange.End() ) {
+					using( new SerializedObjectScope( m_serializedObject ) ) {
+						m_RaycastTarget.boolValue = b;
+					}
+					EditorWindowUtils.RepaintSceneView();
+				}
+			}
+
 			if( UnityTypes.UnityEngine_SpriteRenderer != null && com.GetType() == ( UnityTypes.UnityEngine_SpriteRenderer ) ) {
 				ScopeHorizontal.Begin();
 				if( GUILayout.Button( "#", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "#" );
-					if( GUILayout.Button( "##", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "##" );
-					if( GUILayout.Button( "Set GameObject Name" ) ) EditorHelper.SetGameObjectName( com );
+				if( GUILayout.Button( "##", GUILayout.ExpandWidth( false ) ) ) EditorHelper.SetGameObjectName( com, "##" );
+				if( GUILayout.Button( "Set GameObject Name" ) ) EditorHelper.SetGameObjectName( com );
 				ScopeHorizontal.End();
 
 				var s = com.GetProperty<Sprite>( "sprite" );